Add GH Action to pull wasm artifacts
Introduce a workflow to download the artifacts from the main wgpucube repo and put them in the public/ folder.

This requires a personal access token with to access the artifacts from the other repo. The other repo also needs a second personal access token to trigger this workflow. I might consider simplifying later to a method where the other repo pushes a new commit to this repo which would reduce the need to a single personal access token.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>